supervisor_agent/requirements.txt
langgraph
langgraph-supervisor
langchain-openai
langchain-core
python-dotenv
